backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic>=0.40.0
python-multipart==0.0.9
pydantic==2.6.0
python-dotenv==1.0.1
slowapi==0.1.9
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
